+++ upload/admin/language/en-gb/extension/payment/globalpayments_ucp.php
@@ -28,8 +28,10 @@ $_['label_title']                 = 'Title';
 $_['label_is_production']         = 'Live Mode';
 $_['label_app_id']                = 'Live App Id';
 $_['label_app_key']               = 'Live App Key';
+$_['label_account_name']          = 'Live Account Name';
 $_['label_sandbox_app_id']        = 'Sandbox App Id';
 $_['label_sandbox_app_key']       = 'Sandbox App Key';
+$_['label_sandbox_account_name']  = 'Sandbox Account Name';
 $_['credentials_check']           = 'Credentials check';
 $_['label_debug']                 = 'Enable Logging';
 $_['label_contact_url']           = 'Contact Url';
@@ -52,6 +54,7 @@ $_['help_payment_action']        = 'Choose whether you wish to capture funds imm
 $_['help_allow_card_saving']     = 'Note: to use the card saving feature, you must have multi-use token support enabled on your account. Please contact <a href="mailto:%s?Subject=OpenCart%%20Card%%20Saving%%20Option">support</a> with any questions regarding this option.';
 $_['help_txn_descriptor']        = 'During a Capture or Authorize payment action, this value will be passed along as the transaction-specific descriptor listed on the customer\'s bank account (maxLength: 25).';
 $_['help_txn_descriptor_note']   = 'Please contact <a href="mailto:%s?Subject=OpenCart%%20Transaction%%20Descriptor%%20Option">support</a> with any questions regarding this option.';
+$_['help_account_name']          = 'Specify which account to use when processing a transaction. Default account will be used if this is not specified.';
 
 // Entry
 $_['entry_enabled']                  = 'Enable Gateway';